# api/docs/swagger.yaml
basePath: /
definitions:
  apperror.ErrorCode:
    enum:
    - 400
    - 401
    - 404
    - 500
    type: integer
    x-enum-varnames:
    - ErrCodeBadRequest
    - ErrCodeUnauthorized
    - ErrCodeNotFound
    - ErrCodeInternal
  apperror.Response:
    properties:
      message:
        type: string
      status_code:
        $ref: '#/definitions/apperror.ErrorCode'
    type: object
...
  dto.Block:
    properties:
      hash:
        type: string
      height:
        type: integer
      proposer:
        $ref: '#/definitions/dto.BlockProposer'
      timestamp:
        type: string
      tx_count:
        type: integer
    type: object
  dto.BlockHeightInformativeLatestResponse:
    properties:
      height:
        type: integer
    type: object
  dto.BlockHeightLatestResponse:
    properties:
      height:
        type: integer
    type: object
  dto.BlockInfoResponse:
    properties:
      gas_limit:
        type: integer
      gas_used:
        type: integer
      hash:
        type: string
      height:
        type: integer
      proposer:
        $ref: '#/definitions/dto.BlockProposer'
      timestamp:
        type: string
    type: object
  dto.BlockProposer:
    properties:
      identity:
        type: string
      moniker:
        type: string
      operator_address:
        type: string
    type: object
  dto.BlockTimeAverageResponse:
    properties:
      avg_block_time:
        type: number
    type: object
  dto.BlockTxModel:
    properties:
      address:
        type: string
      hash:
        type: string
      height:
        type: integer
      is_ibc:
        type: boolean
      is_opinit:
        type: boolean
      is_send:
        type: boolean
      messages:
        type: object
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  dto.BlockTxsResponse:
    properties:
      block_txs:
        items:
          $ref: '#/definitions/dto.BlockTxModel'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
  dto.BlocksResponse:
    properties:
      blocks:
        items:
          $ref: '#/definitions/dto.Block'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
...
  dto.PaginationResponse:
    properties:
      next_key:
        type: string
      previous_key:
        type: string
      total:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API service for the Core Indexer project
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Core Indexer API
  version: "1.0"
paths:
...
  /indexer/block/v1/avg_block_time:
    get:
      description: Retrieve the average time taken to mine a block
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlockTimeAverageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get average block time
      tags:
      - Block
  /indexer/block/v1/blocks:
    get:
      consumes:
      - application/json
      description: Retrieve a list of blocks with pagination
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: pagination.offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: pagination.limit
        type: integer
      - default: false
        description: Count total number of transactions
        in: query
        name: pagination.count_total
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlocksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get blocks
      tags:
      - Block
  /indexer/block/v1/blocks/{height}/info:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a block by its height
      parameters:
      - description: Block height
        in: path
        name: height
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlockInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get block info by height
      tags:
      - Block
  /indexer/block/v1/blocks/{height}/txs:
    get:
      consumes:
      - application/json
      description: Retrieve transactions in a block by its height with pagination
      parameters:
      - description: Block height
        in: path
        name: height
        required: true
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: pagination.offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: pagination.limit
        type: integer
      - default: false
        description: Count total number of transactions
        in: query
        name: pagination.count_total
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlockTxsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get transactions in a block by height
      tags:
      - Block
  /indexer/block/v1/latest_block_height:
    get:
      description: Retrieve the latest block height
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlockHeightLatestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get latest block height
      tags:
      - Block
  /indexer/block/v1/latest_informative_block_height:
    get:
      description: Retrieve the latest informative block height
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlockHeightInformativeLatestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get latest informative block height
      tags:
      - Block
...
swagger: "2.0"
...